+++ PSFramework/internal/scripts/postimport.ps1
@@ -5,6 +5,10 @@
 . Import-ModuleFile -Path "$($script:ModuleRoot)\internal\scripts\strings.ps1"
 
 # Initialize the configurations
+foreach ($file in (Get-ChildItem -Path "$($script:ModuleRoot)\internal\configurationschemata\*.ps1"))
+{
+	. Import-ModuleFile -Path $file.FullName
+}
 foreach ($file in (Get-ChildItem -Path "$($script:ModuleRoot)\internal\configurationvalidation\*.ps1"))
 {
 	. Import-ModuleFile -Path $file.FullName
@@ -42,6 +46,9 @@ foreach ($file in (Get-ChildItem -Path "$($script:ModuleRoot)\internal\parameter
 # Import the aliases for PSFramework types
 . Import-ModuleFile -Path "$($script:ModuleRoot)\bin\type-aliases.ps1"
 
+# Import the type extensions used for special types that need to be dynamically calculated
+. Import-ModuleFile -Path "$($script:ModuleRoot)\bin\type-extensions.ps1"
+
 # Register the task engine
 . Import-ModuleFile -Path "$($script:ModuleRoot)\internal\scripts\taskEngine.ps1"
 
@@ -50,6 +57,9 @@ foreach ($file in (Get-ChildItem -Path "$($script:ModuleRoot)\internal\parameter
 
 # Load special variables
 . Import-ModuleFile -Path "$($script:ModuleRoot)\internal\scripts\variables.ps1"
+
+# Load Session Registrations for the Session Container feature
+. Import-ModuleFile -Path "$($script:ModuleRoot)\internal\scripts\sessionRegistration.ps1"
 
 # Load resources for TEPP input completion
 . Import-ModuleFile -Path "$($script:ModuleRoot)\internal\scripts\teppInputResources.ps1"